Fix on the transition of the nsites slot from integer to numeric.

+++ R/methods-GScores.R
@@ -25,9 +25,9 @@ GScores <- function(provider, provider_version, download_url,
 
   nsites <- NA_real_
   if (file.exists(file.path(data_dirpath, "nsites.rds")))
-    nsites <-  as.integer(readRDS(file.path(data_dirpath, "nsites.rds")))
+    nsites <-  as.numeric(readRDS(file.path(data_dirpath, "nsites.rds")))
   else if (file.exists(file.path(data_dirpath, "nov.rds"))) ## temporary during deprecation of MafDb
-    nsites <-  as.integer(readRDS(file.path(data_dirpath, "nov.rds")))
+    nsites <-  as.numeric(readRDS(file.path(data_dirpath, "nov.rds")))
 
   new("GScores", provider=provider,
                  provider_version=provider_version,
